+++ cs-trinity/Wallets/Templates/Messages/FounderSign.cs
@@ -33,7 +33,7 @@ namespace Trinity.Wallets.Templates.Messages
     public class FounderSignBody
     {
         public string AssetType { get; set; }
-        public double Deposit { get; set; }
+        public long Deposit { get; set; }
         public FundingSignTx Founder { get; set; }
         public CommitmentSignTx Commitment { get; set; }
         public RevocableDeliverySignTx RevocableDelivery { get; set; }
